Abbreviate commit hash in prompt during git rebase
The full commit hash is quite long to be included in a prompt.
Given it is unlikely that the hashes from the prompt need long term uniqueness,
we can display the abbreviated hash instead.

If, for some reason, the user wants the full hash, they can set the git's
`core.abbrev` config.

Ths change also allows us to reliably display the commit subject, even in the
edge cases where it is not present in the `.git/rebase-merge/done` file.

https://git-scm.com/docs/git-config#Documentation/git-config.txt-coreabbrev